Enable auto-healing, update to Debian 10
This patch adds an auto-healing policy to automatically re-create the
vault cluster instance if the vault server stops.  One of the nodes in
the instance group is active as per [Vault HA][ha].  The other nodes are
passive and forward requests to the active node.  Two different health
checks are used because passive nodes return non-200 status codes by
default.

In addition, this patch:

 * Update Vault to 1.6.0 by default
 * Update image to Debian 10 by default
 * Defaults to e2-standard-2 instance types, which are less expensive
   and more performant than n1-standard-1.
 * Improves startup (and auto-heal recovery) time by starting the vault
   service as quickly as possible in the startup-script.

[ha]: https://www.vaultproject.io/docs/concepts/ha.html